requirements.txt
# Core dependencies
MainContentExtractor>=0.0.4
beautifulsoup4>=4.12.3
httpx>=0.27.2
langchain==0.3.14
langchain-openai==0.3.1
langchain-anthropic==0.3.3
langchain-ollama==0.2.2
langchain-fireworks>=0.2.6
langchain-aws>=0.2.11
langchain-google-genai==2.0.8
pydantic>=2.10.4
python-dotenv>=1.0.1
requests>=2.32.3
posthog>=3.7.0
playwright>=1.49.0
setuptools>=75.8.0
lmnr[langchain]>=0.4.53

# API dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-socketio>=5.10.0
python-jose[cryptography]>=3.3.0  # For JWT
python-multipart>=0.0.6  # For form data
pydantic-settings>=2.1.0

# Development dependencies
tokencost>=0.1.16
hatch>=1.13.0
build>=1.2.2
pytest>=8.3.3
pytest-asyncio>=0.24.0
httpx>=0.25.0  # For testing FastAPI 